Dev glag

+++ src/parsing/fill_map.c
@@ -6,7 +6,7 @@
 /*   By: glaguyon <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/06/16 19:26:47 by glaguyon          #+#    #+#             */
-/*   Updated: 2024/06/16 20:19:36 by glaguyon         ###   ########.fr       */
+/*   Updated: 2024/06/28 17:52:53 by glaguyon         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -15,8 +15,8 @@
 
 void	fill_map(t_map *map, char *buf)
 {
-	size_t	i;
-	size_t	len;
+	unsigned int	i;
+	unsigned int	len;
 
 	i = 0;
 	while (*buf)
